New file added for testcases
